src/requirements.txt
# Core Django
Django==4.2.20

# Database
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Static files serving
whitenoise==6.6.0

# Web server
gunicorn==21.2.0

# Analytics and data processing
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2

# Forms and UI
django-widget-tweaks==1.5.0

# Data export
openpyxl==3.1.2

# Image processing (for icon generation)
Pillow==11.3.0

# Development tools (remove in production if needed)
django-tailwind==3.6.0
django-browser-reload==1.12.1

# Security (optional but recommended)
django-ratelimit==4.1.0
django-axes==6.1.1

# Caching (optional)
django-redis==5.4.0
redis==5.0.1

# Monitoring (optional)
sentry-sdk==1.40.0